Showed correct comments for tasks (#205, #238)
Issue #205 requested that a comments form be added to the tasks detail modal, which was completed in PR #238.

However, this would only show the comments for the first task viewed. Any subsequent tasks clicked on would still show the comments from the first task.

These commit now allows for the correct comments to be shown for the relevant task.